Extract ClickHouse endpoint validation utility
